// react-frontend/src/api/models/campus-entity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CampusEntity
 */
export interface CampusEntity {
    /**
     * 
     * @type {number}
     * @memberof CampusEntity
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof CampusEntity
     */
    'campus'?: string;
    /**
     * 
     * @type {string}
     * @memberof CampusEntity
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof CampusEntity
     */
    'updatedAt'?: string;
}